+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Unreleased
 - Add `IncrementMetric` to Consume Runtime Parts to allow custom metrics
 - Add `CustomMetrics`
+- Allow Enable/Disable Resource Availability from `ConsumeRuntimeParts`
 
 ## 1.0.0 - 2019-05-29
 - Initial implementation